requirements.txt
# Core testing and tooling
pytest
pytest-cov
black
ruff
isort
python-dotenv
requests
